feat(iso): Create new ISO generator flow using Titanoboa format
